+++ src/main/java/gregtech/api/metatileentity/CommonMetaTileEntity.java
@@ -35,14 +35,14 @@ import gnu.trove.list.TIntList;
 import gnu.trove.list.array.TIntArrayList;
 import gregtech.api.GregTechAPI;
 import gregtech.api.gui.modularui.GTUIInfos;
-import gregtech.api.gui.modularui2.GTGuiTheme;
-import gregtech.api.gui.modularui2.GTGuiThemes;
-import gregtech.api.gui.modularui2.GTGuis;
-import gregtech.api.gui.modularui2.GTModularScreen;
-import gregtech.api.gui.modularui2.MetaTileEntityGuiHandler;
 import gregtech.api.interfaces.metatileentity.IMetaTileEntity;
 import gregtech.api.interfaces.modularui.IAddUIWidgets;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
+import gregtech.api.modularui2.GTGuiTheme;
+import gregtech.api.modularui2.GTGuiThemes;
+import gregtech.api.modularui2.GTGuis;
+import gregtech.api.modularui2.GTModularScreen;
+import gregtech.api.modularui2.MetaTileEntityGuiHandler;
 import gregtech.api.objects.GTItemStack;
 import gregtech.api.util.GTLanguageManager;
 import gregtech.api.util.GTUtility;
